shops/visualize_gradient_descent/requirements.txt
numpy
scipy
scikit-learn
tensorflow
umap-learn
matplotlib
seaborn